data_warehouse/scripts/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
faker>=19.0.0
databricks-sql-connector>=2.0.0
dbt-databricks>=1.7.0


